Revert "Let semaphore guards acquire additional resources"
Using that functionality is begging for a deadlock. What happens when
more than one thread wants to upgrade their resource to have more
values? The semaphore can be completely used up by threads that have
some resources and are waiting to get more. The right way is to release
resources and make a new request.

This reverts commit e0075bb7fbeb68fa1a5b910423e23774545a9cb3.